+++ src/App.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Container, Row, Col } from 'react-bootstrap';
+import Accounts from "./js/components/Accounts";
 import Footer from "./js/components/Footer";
 
 function App() {
@@ -8,7 +9,8 @@ function App() {
     <Container>
       <Row>
         <Col>
-          <h1>Coming Soon</h1>
+          <h1>Accounts</h1>
+          <Accounts />
         </Col>
       </Row>
     </Container>