+++ src/js/reducers/index.js
@@ -1,15 +1,28 @@
-import { LOADED_API_VERSION_SUCCESSFULLY } from "../constants/action-types";
+import { LOADED_API_VERSION_SUCCESSFULLY, LOADED_ACCOUNTS_LIST_SUCCESSFULLY } from "../constants/action-types";
 
 const initialState = {
-  apiVersion: null
+  apiVersion: null,
+  frontendVersion: '0.2.0',
+  accountsList: []
 };
 
 function rootReducer(state = initialState, action) {
 
-  if (action.type === LOADED_API_VERSION_SUCCESSFULLY) {
-    return Object.assign({}, state, {
-      apiVersion: action.payload.api_version
-    });
+  switch (action.type) {
+
+    case LOADED_API_VERSION_SUCCESSFULLY: {
+      return Object.assign({}, state, {
+        apiVersion: action.payload.api_version
+      });
+    }
+
+    case LOADED_ACCOUNTS_LIST_SUCCESSFULLY: {
+      return Object.assign({}, state, {
+        accountsList: action.payload
+      });
+    }
+
+    default: {}
   }
 
   return state;