+++ CHANGELOG.md
@@ -1,2 +1,8 @@
-# 0.1.0
+# 0.2.0
+
+### Add
++ Add Frontend version on store
++ Add Accounts table on frontend
+
+## 0.1.0
 + First version